Completed profile config logic, and rerouted all sensor pages based off Firestore data.
