+++ .gitignore
@@ -24,3 +24,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 test-json.jsonl
+.firebase/
