+++ src/Pages/Landing/Landing.jsx
@@ -1,5 +1,4 @@
 import React, { Component, Fragment } from "react";
-import PropTypes from "prop-types";
 // Material Ui
 import { Paper, Button, Typography, CssBaseline } from "@material-ui/core";
 // Logos
@@ -14,13 +13,6 @@ import { mainTheme } from "../../assets/jss/mainStyle";
 class LandingPage extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      username: "",
-      password: "",
-      error: null,
-      openDialog: false,
-      isRemembered: false
-    };
   }
 
   render() {
@@ -31,7 +23,9 @@ class LandingPage extends Component {
         <CssBaseline />
         <MuiThemeProvider theme={mainTheme}>
           <div className={classes.root}>
-            <img className={classes.logos} src={logos} alt="logos" />
+            <Paper className={classes.paper}>
+              <img className={classes.logos} src={logos} alt="logos" />
+            </Paper>
             <div className={classes.container}>
               <Paper className={classes.paper}>
                 <Typography variant="h5" className={classes.title}>
@@ -45,7 +39,7 @@ class LandingPage extends Component {
                       variant="contained"
                       color="primary"
                     >
-                      Start
+                      START
                     </Button>
                   </Link>
                 </form>
@@ -57,9 +51,5 @@ class LandingPage extends Component {
     );
   }
 }
-
-LandingPage.propTypes = {
-  classes: PropTypes.object.isRequired
-};
 
 export default withStyles(landingStyle)(LandingPage);
